pyproject.toml
[project]
name = "dsgp-brain-tumor-detection"
version = "0.1.0"
description = "Brain Tumor Detection ML/DL Project"
requires-python = ">=3.11,<3.12"

dependencies = [
    "numpy==2.0.2",
    "pandas==2.2.2",
    "scikit-learn==1.6.1",
    "matplotlib==3.10.0",
    "seaborn==0.13.2",
    "PyYAML==6.0.3",
    "streamlit==1.52.2",
    "opencv-python==4.11.0.86",
    "tqdm==4.67.1",
    "Pillow==11.3.0",
    "tensorboard==2.19.0",
    "tensorflow==2.19.0",
    "tensorflow-io-gcs-filesystem==0.31.0",
    "albumentations==2.0.8",
    "scikit-image==0.25.2",
    "timm==1.0.22",
    "einops==0.8.1",
    "shap==0.50.0",
    "hydra-core==1.3.2",
    "wandb==0.23.1",
    "rich==13.9.4",
    "nbformat>=4.5"
]

[dependency-groups]
dev = [
    "pytest==8.4.2",
    "black==25.12.0",
    "flake8==7.3.0"
]

[tool.uv.sources]
torch = { index = "pytorch" }
torchvision = { index = "pytorch" }

[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

# Linting config for flake8 (optional)
[tool.flake8]
max-line-length = 121
